Struct longbridge::quote::QuoteContext

pub struct QuoteContext { /* private fields */ }

Quote context

Implementations

impl QuoteContext

pub async fn try_new(
config: Arc<Config>
) -> Result<(Self, UnboundedReceiver<PushEvent>)>

Create a QuoteContext

pub async fn subscribe<I, T>(
&self,
symbols: I,
sub_types: impl Into<SubFlags>,
is_first_push: bool
) -> Result<()> where
I: IntoIterator<Item = T>,
T: Into<String>,

Subscribe

Reference: https://open.longbridgeapp.com/en/docs/quote/subscribe/subscribe

Examples

use std::sync::Arc;

use longbridge::{
    quote::{QuoteContext, SubFlags},
    Config,
};

let config = Arc::new(Config::from_env()?);
let (ctx, mut receiver) = QuoteContext::try_new(config).await?;

ctx.subscribe(["700.HK", "AAPL.US"], SubFlags::QUOTE, false)
    .await?;
while let Some(msg) = receiver.recv().await {
    println!("{:?}", msg);
}

pub async fn unsubscribe<I, T>(
&self,
symbols: I,
sub_types: impl Into<SubFlags>
) -> Result<()> where
I: IntoIterator<Item = T>,
T: Into<String>,

Unsubscribe

Reference: https://open.longbridgeapp.com/en/docs/quote/subscribe/unsubscribe

Examples

use std::sync::Arc;

use longbridge::{
    quote::{QuoteContext, SubFlags},
    Config,
};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

ctx.subscribe(["700.HK", "AAPL.US"], SubFlags::QUOTE, false)
    .await?;
ctx.unsubscribe(["AAPL.US"], SubFlags::QUOTE).await?;

pub async fn subscribe_candlesticks<T>(
&self,
symbol: T,
period: Period
) -> Result<()> where
T: Into<String>,

Subscribe security candlesticks

Examples

use std::sync::Arc;

use longbridge::{
    quote::{Period, QuoteContext},
    Config,
};

let config = Arc::new(Config::from_env()?);
let (ctx, mut receiver) = QuoteContext::try_new(config).await?;

ctx.subscribe_candlesticks("AAPL.US", Period::OneMinute)
    .await?;
while let Some(msg) = receiver.recv().await {
    println!("{:?}", msg);
}

pub async fn unsubscribe_candlesticks<T>(
&self,
symbol: T,
period: Period
) -> Result<()> where
T: Into<String>,

Unsubscribe security candlesticks

pub async fn subscriptions(&self) -> Result<Vec<Subscription>>

Get subscription information

Examples

use std::sync::Arc;

use longbridge::{
    quote::{QuoteContext, SubFlags},
    Config,
};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

ctx.subscribe(["700.HK", "AAPL.US"], SubFlags::QUOTE, false)
    .await?;
let resp = ctx.subscriptions().await?;
println!("{:?}", resp);

pub async fn static_info<I, T>(
&self,
symbols: I
) -> Result<Vec<SecurityStaticInfo>> where
I: IntoIterator<Item = T>,
T: Into<String>,

Get basic information of securities

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/static

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx
    .static_info(["700.HK", "AAPL.US", "TSLA.US", "NFLX.US"])
    .await?;
println!("{:?}", resp);

pub async fn quote<I, T>(&self, symbols: I) -> Result<Vec<SecurityQuote>> where
I: IntoIterator<Item = T>,
T: Into<String>,

Get quote of securities

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/quote

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx
    .quote(["700.HK", "AAPL.US", "TSLA.US", "NFLX.US"])
    .await?;
println!("{:?}", resp);

pub async fn option_quote<I, T>(&self, symbols: I) -> Result<Vec<OptionQuote>> where
I: IntoIterator<Item = T>,
T: Into<String>,

Get quote of option securities

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/option-quote

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx.option_quote(["AAPL230317P160000.US"]).await?;
println!("{:?}", resp);

pub async fn warrant_quote<I, T>(&self, symbols: I) -> Result<Vec<WarrantQuote>> where
I: IntoIterator<Item = T>,
T: Into<String>,

Get quote of warrant securities

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/warrant-quote

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx.warrant_quote(["21125.HK"]).await?;
println!("{:?}", resp);

pub async fn depth(&self, symbol: impl Into<String>) -> Result<SecurityDepth>

Get security depth

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/depth

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx.depth("700.HK").await?;
println!("{:?}", resp);

pub async fn brokers(&self, symbol: impl Into<String>) -> Result<SecurityBrokers>

Get security brokers

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/brokers

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx.brokers("700.HK").await?;
println!("{:?}", resp);

pub async fn participants(&self) -> Result<Vec<ParticipantInfo>>

Get participants

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/broker-ids

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx.participants().await?;
println!("{:?}", resp);

pub async fn trades(
&self,
symbol: impl Into<String>,
count: usize
) -> Result<Vec<Trade>>

Get security trades

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/trade

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx.trades("700.HK", 10).await?;
println!("{:?}", resp);

pub async fn intraday(
&self,
symbol: impl Into<String>
) -> Result<Vec<IntradayLine>>

Get security intraday lines

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/intraday

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx.intraday("700.HK").await?;
println!("{:?}", resp);

pub async fn candlesticks(
&self,
symbol: impl Into<String>,
period: Period,
count: usize,
adjust_type: AdjustType
) -> Result<Vec<Candlestick>>

Get security candlesticks

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/candlestick

Examples

use std::sync::Arc;

use longbridge::{
    quote::{AdjustType, Period, QuoteContext},
    Config,
};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx
    .candlesticks("700.HK", Period::Day, 10, AdjustType::NoAdjust)
    .await?;
println!("{:?}", resp);

pub async fn option_chain_expiry_date_list(
&self,
symbol: impl Into<String>
) -> Result<Vec<Date>>

Get option chain expiry date list

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/optionchain-date

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx.option_chain_expiry_date_list("AAPL.US").await?;
println!("{:?}", resp);

pub async fn option_chain_info_by_date(
&self,
symbol: impl Into<String>,
expiry_date: Date
) -> Result<Vec<StrikePriceInfo>>

Get option chain info by date

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/optionchain-date-strike

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config};
use time::macros::date;

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx
    .option_chain_info_by_date("AAPL.US", date!(2023 - 01 - 20))
    .await?;
println!("{:?}", resp);

pub async fn warrant_issuers(&self) -> Result<Vec<IssuerInfo>>

Get warrant issuers

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/issuer

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx.warrant_issuers().await?;
println!("{:?}", resp);

pub async fn trading_session(&self) -> Result<Vec<MarketTradingSession>>

Get trading session of the day

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/trade-session

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx.trading_session().await?;
println!("{:?}", resp);

pub async fn trading_days(
&self,
market: Market,
begin: Date,
end: Date
) -> Result<MarketTradingDays>

Get market trading days

The interval must be less than one month, and only the most recent year is supported.

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/trade-day

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config, Market};
use time::macros::date;

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx
    .trading_days(Market::HK, date!(2022 - 01 - 20), date!(2022 - 02 - 20))
    .await?;
println!("{:?}", resp);

pub async fn capital_flow(
&self,
symbol: impl Into<String>
) -> Result<Vec<CapitalFlowLine>>

Get capital flow intraday

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/capital-flow-intraday

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx.capital_flow("700.HK").await?;
println!("{:?}", resp);

pub async fn capital_distribution(
&self,
symbol: impl Into<String>
) -> Result<CapitalDistributionResponse>

Get capital distribution

Reference: https://open.longbridgeapp.com/en/docs/quote/pull/capital-distribution

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx.capital_distribution("700.HK").await?;
println!("{:?}", resp);

pub async fn watch_list(&self) -> Result<Vec<WatchListGroup>>

Get watch list

Reference: https://open.longbridgeapp.com/en/docs/quote/individual/watchlist_groups

Examples

use std::sync::Arc;

use longbridge::{quote::QuoteContext, Config};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

let resp = ctx.watch_list().await?;
println!("{:?}", resp);

pub async fn realtime_quote<I, T>(
&self,
symbols: I
) -> Result<Vec<RealtimeQuote>> where
I: IntoIterator<Item = T>,
T: Into<String>,

Get real-time quotes

Get real-time quotes of the subscribed symbols, it always returns the data in the local storage.

Examples

use std::{sync::Arc, time::Duration};

use longbridge::{
    quote::{QuoteContext, SubFlags},
    Config,
};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

ctx.subscribe(["700.HK", "AAPL.US"], SubFlags::QUOTE, true)
    .await?;
tokio::time::sleep(Duration::from_secs(5)).await;

let resp = ctx.realtime_quote(["700.HK", "AAPL.US"]).await?;
println!("{:?}", resp);

pub async fn realtime_depth(
&self,
symbol: impl Into<String>
) -> Result<SecurityDepth>

Get real-time depth

Get real-time depth of the subscribed symbols, it always returns the data in the local storage.

Examples

use std::{sync::Arc, time::Duration};

use longbridge::{
    quote::{QuoteContext, SubFlags},
    Config,
};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

ctx.subscribe(["700.HK", "AAPL.US"], SubFlags::DEPTH, true)
    .await?;
tokio::time::sleep(Duration::from_secs(5)).await;

let resp = ctx.realtime_depth("700.HK").await?;
println!("{:?}", resp);

pub async fn realtime_trades(
&self,
symbol: impl Into<String>,
count: usize
) -> Result<Vec<Trade>>

Get real-time trades

Get real-time trades of the subscribed symbols, it always returns the data in the local storage.

Examples

use std::{sync::Arc, time::Duration};

use longbridge::{
    quote::{QuoteContext, SubFlags},
    Config,
};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

ctx.subscribe(["700.HK", "AAPL.US"], SubFlags::TRADE, false)
    .await?;
tokio::time::sleep(Duration::from_secs(5)).await;

let resp = ctx.realtime_trades("700.HK", 10).await?;
println!("{:?}", resp);

pub async fn realtime_brokers(
&self,
symbol: impl Into<String>
) -> Result<SecurityBrokers>

Get real-time broker queue

Get real-time broker queue of the subscribed symbols, it always returns the data in the local storage.

Examples

use std::{sync::Arc, time::Duration};

use longbridge::{
    quote::{QuoteContext, SubFlags},
    Config,
};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

ctx.subscribe(["700.HK", "AAPL.US"], SubFlags::BROKER, true)
    .await?;
tokio::time::sleep(Duration::from_secs(5)).await;

let resp = ctx.realtime_brokers("700.HK").await?;
println!("{:?}", resp);

pub async fn realtime_candlesticks(
&self,
symbol: impl Into<String>,
period: Period,
count: usize
) -> Result<Vec<Candlestick>>

Get real-time candlesticks

Get real-time candlesticks of the subscribed symbols, it always returns the data in the local storage.

Examples

use std::{sync::Arc, time::Duration};

use longbridge::{
    quote::{Period, QuoteContext},
    Config,
};

let config = Arc::new(Config::from_env()?);
let (ctx, _) = QuoteContext::try_new(config).await?;

ctx.subscribe_candlesticks("AAPL.US", Period::OneMinute)
    .await?;
tokio::time::sleep(Duration::from_secs(5)).await;

let resp = ctx
    .realtime_candlesticks("AAPL.US", Period::OneMinute, 10)
    .await?;
println!("{:?}", resp);